+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [7.2.0] - 2024-01-??
 
 - Fixed: Bug where tooltips did not show uncollected item names in the autotracker.
+- Fixed: The import in a multiworld session is blocked if it contains an unsupported game.
 - Changed: Update to the Database Video Directory site to eliminate lag and add modern styling.
 - Changed: Autotracker tooltips now display text in black instead of gray.